+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ## ScottPlot 5.0.12 (in development)
 * Axes: Improved automatic axis scaling for plots containing 1D plottables (#3132)
 * Coordinates: Added `AreReal` property to confirm `X` and `Y` are finite
+* Crosshair: Added `X` and `Y` properties to compliment `Position`
 
 ## ScottPlot 5.0.11-beta
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2023-12-29_